Update README.md
Welcome to Stories Store, your gateway to a virtual world of literary treasures. Stories Store is not just an online bookstore; it's an immersive experience crafted with HTML, CSS, and JavaScript. Explore my GitHub repository to witness the fusion of storytelling. From the artful design in HTML and CSS to the interactive features powered by JavaScript, Stories Store showcases the magic of front-end web development in bringing books to life. Dive into my code and discover how I've created a seamless online book-shopping experience. It's not just a website; it's a testament to my passion for coding and love for literature. Join me in this digital journey through the captivating world of books, all at Stories Store on my GitHub account.